Evaluate dag
